aider: Define a job for the `reddit_submissions` asset in the `__init__.py` file.
# Aider chat conversation:

USER: okay. are there any changes we need to make to the code in order to get the asset to materialize, like do we need to run a job or anything?/add anarchists-cookbook/anarchists_cookbook/__init__.py

ASSISTANT: To materialize the `reddit_submissions` asset, you need to define a job that includes this asset. A job in Dagster is a pipeline of execution for one or more assets. You can define a job in the `__init__.py` file that includes the `reddit_submissions` asset. This job can then be executed to materialize the asset.

Here is the diff for the changes to the `__init__.py` file to define a job for the `reddit_submissions` asset:

```diff
--- anarchists-cookbook/anarchists_cookbook/__init__.py
+++ anarchists-cookbook/anarchists_cookbook/__init__.py
@@ ... @@
 from .assets.reddit_posts_comments_df import reddit_submissions

 defs = Definitions(
     assets=load_assets_from_modules([assets]),  # Adjust as necessary
     resources={
         "reddit_client": reddit_client
         # Include other resources as before
     },
 )
+# Define a job for the reddit_submissions asset
+reddit_submissions_job = define_asset_job(
+    "reddit_submissions_job",
+    selection=AssetSelection.assets(reddit_submissions),
+)
+
 # TODO: import redditdatagenerator from .resources, see tutorial/__init__.py
```

This change adds a job called `reddit_submissions_job` that specifically targets the `reddit_submissions` asset. Once this job is defined, you can execute it using Dagster's command-line interface or through the Dagster UI to materialize the asset.